fix: clean up test
